pyproject.toml
[tool.poetry]
name = "profile"
version = "0.1.0"
description = ""
authors = ["Pine Mizune <pinemz@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.5.0"
pyyaml = "^6.0"
pelican = {extras = ["markdown"], version = "^4.8.0"}
jinja2 = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
